+++ NYTPhotoViewer/NYTPhotosViewController.h
@@ -58,10 +58,8 @@ extern NSString * const NYTPhotosViewControllerDidDismissNotification;
  *  The data source underlying this PhotosViewController.
  *
  *  After setting a new data source, you must call `-reloadPhotosAnimated:`.
- *
- *  The data source is retained (unlike UITableView's data source).
  */
-@property (nonatomic) id <NYTPhotoViewerDataSource> dataSource;
+@property (nonatomic, weak, nullable) id <NYTPhotoViewerDataSource> dataSource;
 
 /**
  *  The object conforming to `NYTPhoto` that is currently being displayed by the `pageViewController`.